Move each setting to separate boltdb bucket key
This is better because in future I can just add or remove those keys freely, before this making changes to settings would break things. This will ignore old settings, so when updating to this commit it will update old settings to default

+++ menu.go
@@ -44,7 +44,7 @@ type menu struct {
 	settingsState string
 	// DEFAULT, LIGHT, DARK, MONO
 	settingsThemeState   string
-	settingsMaxScrolloff int
+	settingsMaxScrolloff uint64
 }
 
 func (a *app) createMenu() {
@@ -371,7 +371,7 @@ func (a *app) drawMenuSettings() {
 	}
 	currStart += len(monoThemeStr) + 1
 
-	maxScrolloffStr := strconv.Itoa(a.menu.settingsMaxScrolloff)
+	maxScrolloffStr := strconv.Itoa(int(a.menu.settingsMaxScrolloff))
 	a.setContentString(0, 4, a.defStyle, maxScrolloffStr)
 }
